requirements.txt
Pillow>=10.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
facenet-pytorch>=2.5.0
torch>=2.0.0
torchvision>=0.15.0

